requirements.txt
gradio
curl_cffi
selectolax
smolagents
python-dotenv